+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/CircuitBreakerFactory.java
@@ -1,0 +1,61 @@
+package com.mageddo.dnsproxyserver.solver.remote;
+
+import com.mageddo.commons.circuitbreaker.CircuitCheckException;
+import com.mageddo.dnsproxyserver.config.application.ConfigService;
+import com.mageddo.dnsproxyserver.solver.application.CircuitBreakerStateMapper;
+import com.mageddo.dnsproxyserver.solver.dataprovider.SolverRemoteConsistencyGuaranteeDAO;
+import dev.failsafe.CircuitBreaker;
+import dev.failsafe.event.CircuitBreakerStateChangedEvent;
+import dev.failsafe.event.EventListener;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+import javax.inject.Inject;
+import java.net.InetSocketAddress;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
+@Slf4j
+@RequiredArgsConstructor(onConstructor = @__({@Inject}))
+public class CircuitBreakerFactory {
+
+  private final Map<InetSocketAddress, CircuitBreaker<Result>> circuitBreakerMap = new ConcurrentHashMap<>();
+  private final ConfigService configService;
+  private final SolverRemoteConsistencyGuaranteeDAO consistencyGuaranteeDAO;
+
+  public CircuitBreaker<Result> createCircuitBreakerFor(InetSocketAddress address) {
+    final var config = this.findCircuitBreakerConfig();
+    return this.circuitBreakerMap.computeIfAbsent(address, addr -> buildCircuitBreaker(addr, config));
+  }
+
+  CircuitBreaker<Result> buildCircuitBreaker(
+    InetSocketAddress address, com.mageddo.dnsproxyserver.config.CircuitBreaker config
+  ) {
+    return CircuitBreaker.<Result>builder()
+      .handle(CircuitCheckException.class)
+      .withFailureThreshold(config.getFailureThreshold(), config.getFailureThresholdCapacity())
+      .withSuccessThreshold(config.getSuccessThreshold())
+      .withDelay(config.getTestDelay())
+//      .onClose(bindStateChangeEvent("closed", address))
+//      .onOpen(bindStateChangeEvent("open", address))
+//      .onHalfOpen(bindStateChangeEvent("half-open", address))
+      .build();
+  }
+
+  EventListener<CircuitBreakerStateChangedEvent> bindStateChangeEvent(String actualStateName, InetSocketAddress address) {
+    return event -> {
+      final var previousStateName = CircuitBreakerStateMapper.toStateNameFrom(event);
+      this.consistencyGuaranteeDAO.flushCachesFromCircuitBreakerStateChange();
+      log.debug(
+        "status=clearedCache, address={}, previousStateName={}, actualStateName={}",
+        address, previousStateName, actualStateName
+      );
+    };
+  }
+
+  com.mageddo.dnsproxyserver.config.CircuitBreaker findCircuitBreakerConfig() {
+    return this.configService.findCurrentConfig()
+      .getSolverRemote()
+      .getCircuitBreaker();
+  }
+}
