+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/Request.java
@@ -1,0 +1,54 @@
+package com.mageddo.dnsproxyserver.solver.remote;
+
+import com.mageddo.dnsproxyserver.solver.Resolver;
+import com.mageddo.net.IpAddr;
+import com.mageddo.net.IpAddrs;
+import lombok.Builder;
+import lombok.NonNull;
+import lombok.Value;
+import org.apache.commons.lang3.time.StopWatch;
+import org.xbill.DNS.Message;
+
+import java.net.InetSocketAddress;
+import java.util.concurrent.CompletableFuture;
+
+@Value
+@Builder
+public class Request {
+
+  @NonNull
+  Message query;
+
+  @NonNull
+  Resolver resolver;
+
+  @NonNull
+  Integer resolverIndex;
+
+  @NonNull
+  StopWatch stopWatch;
+
+  public IpAddr getResolverAddr() {
+    return IpAddrs.from(this.getResolverAddress());
+  }
+
+  public InetSocketAddress getResolverAddress() {
+    return this.getResolver().getAddress();
+  }
+
+  public void splitStopWatch() {
+    this.stopWatch.split();
+  }
+
+  public CompletableFuture<Message> sendQueryAsyncToResolver() {
+    return this.resolver.sendAsync(this.query).toCompletableFuture();
+  }
+
+  public long getElapsedTimeInMs() {
+    return this.stopWatch.getTime() - this.stopWatch.getSplitTime();
+  }
+
+  public long getTime() {
+    return this.stopWatch.getTime();
+  }
+}
